+++ clients/python-client/pyproject.toml
@@ -16,7 +16,6 @@ dependencies = [
     "pydantic>=2.0.0",
     "pytest-asyncio>=0.21.0",
     "python-dotenv>=0.19.0",
-    "requests>=2.28.0",
 ]
 
 [project.optional-dependencies]
