+++ clients/python-client/src/chunkr_ai/__init__.py
@@ -1,4 +1,3 @@
 from .api.chunkr import Chunkr
-from .api.chunkr_async import ChunkrAsync
 
-__all__ = ["Chunkr", "ChunkrAsync"]
+__all__ = ["Chunkr"]
