+++ clients/python-client/src/chunkr_ai/api/chunkr_base.py
@@ -1,13 +1,16 @@
 from .config import Configuration
-from .task import TaskResponse
-from .task_async import TaskResponseAsync
+from .task_response import TaskResponse
 from .auth import HeadersMixin
 from abc import abstractmethod
 from dotenv import load_dotenv
+import httpx
+import io
+import json
 import os
 from pathlib import Path
 from PIL import Image
-from typing import BinaryIO, Union
+import requests
+from typing import BinaryIO, Tuple, Union
 
 
 class ChunkrBase(HeadersMixin):
@@ -23,13 +26,138 @@ class ChunkrBase(HeadersMixin):
             )
 
         self.url = self.url.rstrip("/")
+        self._client = httpx.AsyncClient()
+
+    def _prepare_file(
+        self, file: Union[str, Path, BinaryIO, Image.Image]
+    ) -> Tuple[str, BinaryIO]:
+        """Convert various file types into a tuple of (filename, file-like object).
+
+        Args:
+            file: Input file, can be:
+                - String or Path to a file
+                - URL string starting with http:// or https://
+                - Base64 string
+                - Opened binary file (mode='rb')
+                - PIL/Pillow Image object
+
+        Returns:
+            Tuple[str, BinaryIO]: (filename, file-like object) ready for upload
+
+        Raises:
+            FileNotFoundError: If the file path doesn't exist
+            TypeError: If the file type is not supported
+            ValueError: If the URL is invalid or unreachable
+            ValueError: If the MIME type is unsupported
+        """
+        # Handle URLs
+        if isinstance(file, str) and (
+            file.startswith("http://") or file.startswith("https://")
+        ):
+            response = requests.get(file)
+            response.raise_for_status()
+            file_obj = io.BytesIO(response.content)
+            filename = Path(file.split("/")[-1]).name or "downloaded_file"
+            return filename, file_obj
+
+        # Handle base64 strings
+        if isinstance(file, str) and "," in file and ";base64," in file:
+            try:
+                # Split header and data
+                header, base64_data = file.split(",", 1)
+                import base64
+
+                file_bytes = base64.b64decode(base64_data)
+                file_obj = io.BytesIO(file_bytes)
+
+                # Try to determine format from header
+                format = "bin"
+                mime_type = header.split(":")[-1].split(";")[0].lower()
+
+                # Map MIME types to file extensions
+                mime_to_ext = {
+                    "application/pdf": "pdf",
+                    "application/vnd.openxmlformats-officedocument.wordprocessingml.document": "docx",
+                    "application/msword": "doc",
+                    "application/vnd.openxmlformats-officedocument.presentationml.presentation": "pptx",
+                    "application/vnd.ms-powerpoint": "ppt",
+                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": "xlsx",
+                    "application/vnd.ms-excel": "xls",
+                    "image/jpeg": "jpg",
+                    "image/png": "png",
+                    "image/jpg": "jpg",
+                }
+
+                if mime_type in mime_to_ext:
+                    format = mime_to_ext[mime_type]
+                else:
+                    raise ValueError(f"Unsupported MIME type: {mime_type}")
+
+                return f"file.{format}", file_obj
+            except Exception as e:
+                raise ValueError(f"Invalid base64 string: {str(e)}")
+
+        # Handle file paths
+        if isinstance(file, (str, Path)):
+            path = Path(file).resolve()
+            if not path.exists():
+                raise FileNotFoundError(f"File not found: {file}")
+            return path.name, open(path, "rb")
+
+        # Handle PIL Images
+        if isinstance(file, Image.Image):
+            img_byte_arr = io.BytesIO()
+            format = file.format or "PNG"
+            file.save(img_byte_arr, format=format)
+            img_byte_arr.seek(0)
+            return f"image.{format.lower()}", img_byte_arr
+
+        # Handle file-like objects
+        if hasattr(file, "read") and hasattr(file, "seek"):
+            # Try to get the filename from the file object if possible
+            name = (
+                getattr(file, "name", "document")
+                if hasattr(file, "name")
+                else "document"
+            )
+            return Path(name).name, file
+
+        raise TypeError(f"Unsupported file type: {type(file)}")
+
+    def _prepare_upload_data(
+        self,
+        file: Union[str, Path, BinaryIO, Image.Image],
+        config: Configuration = None,
+    ) -> Tuple[dict, dict]:
+        """Prepare files and data dictionaries for upload.
+
+        Args:
+            file: The file to upload
+            config: Optional configuration settings
+
+        Returns:
+            Tuple[dict, dict]: (files dict, data dict) ready for upload
+        """
+        filename, file_obj = self._prepare_file(file)
+        files = {"file": (filename, file_obj)}
+        data = {}
+
+        if config:
+            config_dict = config.model_dump(mode="json", exclude_none=True)
+            for key, value in config_dict.items():
+                if isinstance(value, dict):
+                    files[key] = (None, json.dumps(value), "application/json")
+                else:
+                    data[key] = value
+
+        return files, data
 
     @abstractmethod
     def upload(
         self,
         file: Union[str, Path, BinaryIO, Image.Image],
         config: Configuration = None,
-    ) -> Union[TaskResponse, TaskResponseAsync]:
+    ) -> TaskResponse:
         """Upload a file and wait for processing to complete.
 
         Args:
@@ -64,7 +192,7 @@ class ChunkrBase(HeadersMixin):
     @abstractmethod
     def update(
         self, task_id: str, config: Configuration
-    ) -> Union[TaskResponse, TaskResponseAsync]:
+    ) -> TaskResponse:
         """Update a task by its ID and wait for processing to complete.
 
         Args:
@@ -81,7 +209,7 @@ class ChunkrBase(HeadersMixin):
         self,
         file: Union[str, Path, BinaryIO, Image.Image],
         config: Configuration = None,
-    ) -> Union[TaskResponse, TaskResponseAsync]:
+    ) -> TaskResponse:
         """Upload a file for processing and immediately return the task response. It will not wait for processing to complete. To wait for the full processing to complete, use `task.poll()`.
 
         Args:
@@ -117,7 +245,7 @@ class ChunkrBase(HeadersMixin):
     @abstractmethod
     def update_task(
         self, task_id: str, config: Configuration
-    ) -> Union[TaskResponse, TaskResponseAsync]:
+    ) -> TaskResponse:
         """Update a task by its ID and immediately return the task response. It will not wait for processing to complete. To wait for the full processing to complete, use `task.poll()`.
 
         Args:
@@ -130,7 +258,7 @@ class ChunkrBase(HeadersMixin):
         pass
 
     @abstractmethod
-    def get_task(self, task_id: str) -> Union[TaskResponse, TaskResponseAsync]:
+    def get_task(self, task_id: str) -> TaskResponse:
         """Get a task response by its ID.
 
         Args:
@@ -157,4 +285,10 @@ class ChunkrBase(HeadersMixin):
         Args:
             task_id: The ID of the task to cancel
         """
+        pass
+
+    @abstractmethod
+    def close(self) -> None:
+        """Close the client connection. 
+        This should be called when you're done using the client to properly clean up resources."""
         pass
