+++ clients/python-client/src/chunkr_ai/models.py
@@ -4,25 +4,20 @@ from .api.config import (
     ChunkProcessing,
     Configuration,
     CroppingStrategy,
-    ExtractedJson,
     GenerationStrategy,
     GenerationConfig,
-    JsonSchema,
     Model,
     OCRResult,
     OcrStrategy,
     OutputResponse,
-    Property,
     Segment,
     SegmentProcessing,
     SegmentType,
     SegmentationStrategy,
     Status,
-    PipelineType,
+    Pipeline,
 )
-
-from .api.task import TaskResponse
-from .api.task_async import TaskResponseAsync
+from .api.task_response import TaskResponse
 
 __all__ = [
     "BoundingBox",
@@ -30,21 +25,17 @@ __all__ = [
     "ChunkProcessing",
     "Configuration",
     "CroppingStrategy",
-    "ExtractedJson",
     "GenerationConfig",
     "GenerationStrategy",
-    "JsonSchema",
     "Model",
     "OCRResult",
     "OcrStrategy",
     "OutputResponse",
-    "Property",
     "Segment",
     "SegmentProcessing",
     "SegmentType",
     "SegmentationStrategy",
     "Status",
     "TaskResponse",
-    "TaskResponseAsync",
-    "PipelineType",
+    "Pipeline",
 ]
